feat(pack-up-module): [2/n] mock api
exposing the mock API when the third-party lib the API depends on is
unavailable

test result:
```
>>> from lazymodule import BertEmbeddings
>>> BertEmbeddings()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/mnt/d/github/lazymodule/src/lazymodule/utils/dummy_pt.py", line 6, in __init__
    require_torch(self)
  File "/mnt/d/github/lazymodule/src/lazymodule/utils/import_utils.py", line 28, in require_torch
    raise ImportError(PYTORCH_IMPORT_ERROR.format(name))
ImportError:
BertEmbeddings requires the PyTorch library but it was not found in your environment. Checkout the instructions on
the installation page: https://pytorch.org/get-started/locally/ and follow the ones that match your
environment.

>>>
```